fix: request, response, result 등 네이밍 변경
