+++ delivery-service/src/main/java/com/firstlogistics/deliverservice/application/DeliveryCommandService.java
@@ -1,7 +1,7 @@
 package com.firstlogistics.deliverservice.application;
 
 import com.firstlogistics.deliverservice.application.dto.command.CreateDeliveryCommand;
-import com.firstlogistics.deliverservice.application.dto.result.DeliveryCreateResult;
+import com.firstlogistics.deliverservice.application.dto.result.CreateDeliveryResult;
 import com.firstlogistics.deliverservice.application.publisher.DeliveryEventPublisher;
 import com.firstlogistics.deliverservice.domain.entity.Delivery;
 import com.firstlogistics.deliverservice.domain.entity.DeliveryRoute;
@@ -37,7 +37,7 @@ public class DeliveryCommandService {
 	private final UserPort userPort;
 	private final DeliveryEventPublisher deliveryEventPublisher;
 
-	public DeliveryCreateResult createDelivery(
+	public CreateDeliveryResult createDelivery(
 			CreateDeliveryCommand command,
 			CompanyResponse supplierCompany,
 			CompanyResponse receiverCompany,
@@ -140,7 +140,7 @@ public class DeliveryCommandService {
 		);
 		deliveryEventPublisher.publishedDeliveryCreated(deliveryCreatedEvent);
 
-		return DeliveryCreateResult.from(savedDelivery, command, receiver.name());
+		return CreateDeliveryResult.from(savedDelivery, command, receiver.name());
 	}
 
 	private DeliveryCreatedEvent buildDeliveryCreatedEvent(
