+++ delivery-service/src/main/java/com/firstlogistics/deliverservice/presentation/dto/request/CreateDeliveryRequest.java
@@ -12,7 +12,7 @@ import java.time.LocalDateTime;
 import java.util.List;
 import java.util.UUID;
 
-public record DeliveryCreateRequest(
+public record CreateDeliveryRequest(
 	@NotNull UUID orderId,
 	@NotNull LocalDateTime orderedAt,
 	@NotNull LocalDateTime orderDueDate,
